// bunq-budget-gaurd/src/lib/bunq/types.ts
export interface UserPerson {
  id: number;
  created: string;
  updated: string;
  status: string;
  sub_status: string;
  public_uuid: string;
  display_name: string;
  public_nick_name: string;
  language: string;
  region: string;
  session_timeout: number;
  daily_limit_without_confirmation_login: {
    currency: string;
    value: string;
  };
  relations: any[];
  alias: UserAlias[];
  avatar: UserAvatar;
  tax_resident: null;
  notification_filters: NotificationFilter[];
  address_main: Address;
  address_postal: Address;
  address_shipping: null;
  first_name: string;
  middle_name: string;
  last_name: string;
  legal_name: string;
  date_of_birth: string;
  place_of_birth: string;
  country_of_birth: string;
  nationality: string;
  all_nationality: string[];
  gender: string;
  version_terms_of_service: string;
  deny_reason: null;
  document_issuing_authority: null;
  document_expiry_date: null;
  document_status: string;
  is_primary_document: boolean;
  customer: Customer;
  customer_limit: CustomerLimit;
  billing_contract: BillingContract[];
  pack_membership: null;
  premium_trial: null;
}
// ...
export interface ClientKey {
  private_key_client: string;
  public_key_client: string;
  public_key_client_fmt: string;
}

// Add new interfaces for bunq installation response
export interface BunqInstallationResponse {
  Response: BunqInstallationResponseItem[];
}

export interface BunqInstallationResponseItem {
  Id?: {
    id: number;
  };
  Token?: {
    id: number;
    created: string;
    updated: string;
    token: string;
  };
  ServerPublicKey?: {
    server_public_key: string;
  };
}

export type GenerateSignature = (body: string, privateKey: string) => string;

// Add new interfaces for bunq session response
export interface BunqSessionResponse {
  Response: BunqSessionResponseItem[];
}

export interface BunqSessionResponseItem {
  Id?: {
    id: number;
  };
  Token?: {
    id: number;
    created: string;
    updated: string;
    token: string;
  };
  UserPerson?: UserPerson;
}

export interface BunqAccountResponse {
  Response: BunqAccountResponseItem[];
}

export interface BunqAccountResponseItem {
  Id: {
    id: number;
  };
}